Map Update!
MAP UPDATE (III) (By Juani2400)

- Modified: The Security Arrivals post has been deleted and turned into
a 'Construction Area' until something fits in there.
- Added: A Space Cleaner spray has been added to Security! You can found
it in the medical area.
- Modified: The Starboard Emergency storage has been turned into a
Security post!
- Modified: The Chapel doors have been moved to the front of it, leading
to the main Hallway.
- Added: Science has got a new Break Room!
- Modified: The Engineering Emergency Storage walls have been changed
into windows, to make it more visible to Engineers.
- Tweak: The engine's pipes have been moved a bit, to avoid overlapping.
- Added: A new scrubbers emptying system has been added in Engineering.
- Added: CE's office now has some privacy shutters.
- Added: A new engine telescreen now fills the space that was left after
deleting the button from the office.
- Fixed: Two cameras were missing on the Engine.
- Added: A new Training Facility has been added to Security!
- Modified: The Aft Port Solar has been made a bit bigger to accomodate
the increasing need of power in the station in case the main power
sourse fails.
- Modified: The pool now has a better texture.
- Added: The shuttle now have a cargo compartment! Now you can store
your crates and mechs in there, thus not congesting the shuttle.
- Modified: The Bunker now has a bigger capacity, and it has been
redesigned.

Also, I'm planning the following for the next update:
- Doing some changes in Medbay to give more use to the Recovery Room.
This may include moving rooms, creation of new things, deletions...
- The testing for the head dormitories will end in the next update,
which will conclude deleting it due lack of usage.
- I'd like to make some new changes to Cargo.
- The Medbay will get showers and bathrooms, as per request. This is
useful for long-staying patients.
- I'm planning on dividing the Bar and the Kitchen. What do you think of
that? Please post your answers in the "Map suggestions" post of mine.
- As always, I'm open for suggestions. Please enjoy the update, and
don't hesitate to say something you don't like or something you'd like
to see in that post (^) above mentioned.

Kind regards, Juani2400.